Refactor `bson2` encoding/decoding
