+++ internal/bson2/array.go
@@ -23,16 +23,6 @@ import (
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 )
 
-// RawArray represents a BSON array in the binary encoded form.
-//
-// It generally references a part of a larger slice, not a copy.
-type RawArray []byte
-
-// LogValue implements slog.LogValuer interface.
-func (arr *RawArray) LogValue() slog.Value {
-	return slogValue(arr)
-}
-
 // Array represents a BSON array in the (partially) decoded form.
 type Array struct {
 	elements []any
